+++ src/components/portal/portal.js
@@ -2,15 +2,34 @@ import React from 'react';
 import ReactPortal from 'react-portal';
 import PropTypes from 'prop-types';
 
-const Portal = ({ open, children }) => (
-  <ReactPortal isOpened={ open }>
-    { children }
-  </ReactPortal>
-);
+class Portal extends React.Component {
+  static propTypes = {
+    /**
+     * Determine if the portal is visible or not
+     *
+     * @property open
+     * @type {Boolean}
+    */
+    open: PropTypes.bool,
 
-Portal.propTypes = {
-  open: PropTypes.bool,
-  children: PropTypes.node
-};
+    /**
+     * The content of the portal.
+     *
+     * @property children
+     * @type {Node}
+     */
+    children: PropTypes.node
+  }
+
+  render() {
+    return (
+      <ReactPortal isOpened={ this.props.open }>
+        <div ref={ (node) => { this._portal = node; } }>
+          { this.props.children }
+        </div>
+      </ReactPortal>
+    );
+  }
+}
 
 export default Portal;